refactor(dreamer): improve code formatting and line breaks
